LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/stale-while-revalidate/stale-script-expected.txt


FAIL Cache returns stale resource assert_equals: (revalidated version loaded) expected 2 but got 1
